requirements.txt
google.cloud
google-cloud-storage
